cleaned up axios calls in SearchBar - created new getAutoComplete and getSearchResults components
